#87 trim comments from input for each parser input and convert style parser to object
